refactor: extract layout and dashboard components
