+++ client/src/hooks/useGreeting.js
@@ -1,0 +1,10 @@
+import { useMemo } from 'react';
+
+export default function useGreeting() {
+  return useMemo(() => {
+    const hour = new Date().getHours();
+    if (hour < 12) return 'Good morning';
+    if (hour < 18) return 'Good afternoon';
+    return 'Good evening';
+  }, []);
+}
